Chore: [BADA-378] CSS 조정 #272
