+++ src/features/rental/search/page/SearchPosPage.tsx
@@ -1,216 +1,68 @@
 'use client';
 
-import React, { useCallback, useEffect, useMemo, useRef } from 'react';
+import { lazy, Suspense } from 'react';
 
-import { useRouter } from 'next/navigation';
-
-import { useQueryClient } from '@tanstack/react-query';
-
-import { updateAddressUsageTime } from '@/features/rental/search/api/apis';
-import { useCreateAddressHistory } from '@/features/rental/search/hook/useAddressHooks';
-import { useDeleteAddressHistory } from '@/features/rental/search/hook/useDeleteAddressHooks';
-import { useGetAddressHistoryInfinite } from '@/features/rental/search/hook/useGetAddressHistoryInfiniteHooks';
-import { useSearchPlaces } from '@/features/rental/search/hook/useSearchPlacesHooks';
-import useThrottledScroll from '@/features/rental/search/hook/useThrottledScrollHooks';
-import AddressHistoryList from '@/features/rental/search/ui/AddressHistoryList';
-import SearchInputField from '@/features/rental/search/ui/SearchInputField';
-import SearchResults from '@/features/rental/search/ui/SearchResults';
-import { isLoggedIn } from '@/features/rental/search/utils/auth/isLoggedIn';
+import { useSearchPos } from '@/features/rental/search/hook/useSearchPosHooks';
 import { BaseLayout } from '@/shared/ui/BaseLayout';
 import { Header_Detail } from '@/shared/ui/Header_Detail/Header_Detail';
 
-import type { PlaceSearchResult } from '@/features/rental/search/utils/address/searchPlaces';
+// Lazy loaded components for performance optimization
+const SearchInputField = lazy(() =>
+  import('@/features/rental/search/ui/SearchInputField').then((module) => ({
+    default: module.default,
+  })),
+);
+
+const SearchResults = lazy(() =>
+  import('@/features/rental/search/ui/SearchResults').then((module) => ({
+    default: module.default,
+  })),
+);
+
+const AddressHistoryList = lazy(() =>
+  import('@/features/rental/search/ui/AddressHistoryList').then((module) => ({
+    default: module.default,
+  })),
+);
+
+// Loading fallback components
+const SearchInputLoadingFallback = () => (
+  <div className="w-full h-16 bg-[var(--gray-light)] rounded-lg animate-pulse mb-4" />
+);
+
+const SearchResultsLoadingFallback = () => (
+  <div className="space-y-4">
+    {[...Array(5)].map((_, index) => (
+      <div key={index} className="h-16 bg-[var(--gray-light)] rounded-lg animate-pulse" />
+    ))}
+  </div>
+);
+
+const AddressHistoryLoadingFallback = () => (
+  <div className="space-y-3">
+    {[...Array(3)].map((_, index) => (
+      <div key={index} className="h-12 bg-[var(--gray-light)] rounded-lg animate-pulse" />
+    ))}
+  </div>
+);
 
 export default function SearchPosPage() {
-  const queryClient = useQueryClient();
-  const router = useRouter();
-  const createAddressMutation = useCreateAddressHistory();
-  const deleteAddressMutation = useDeleteAddressHistory();
-  const scrollRef = useRef<HTMLDivElement>(null);
-
-  // 로그인 상태에 따라 정렬 결정
-  const [isAuthenticated, setIsAuthenticated] = React.useState<boolean>(false);
-  const sort = isAuthenticated ? 'createdAt,desc' : 'lastUsed,desc';
-
-  // 로그인 상태 감지
-  useEffect(() => {
-    setIsAuthenticated(isLoggedIn());
-  }, []);
-
-  const {
-    data: addressHistoryInfinite,
-    fetchNextPage,
-    hasNextPage,
-    isFetchingNextPage,
-    refetch,
-  } = useGetAddressHistoryInfinite(5, sort);
-
-  // 키워드 검색 훅
   const {
     keyword,
-    setKeyword,
     searchResults,
-    isLoading: isSearchLoading,
+    addressHistoryInfinite,
+    hasAddressHistory,
+    isSearchLoading,
     isLoadingMore,
+    isFetchingNextPage,
     hasNext,
+    handleSelectPlace,
+    handleAddressHistoryClick,
+    handleDeleteAddress,
+    handleKeywordChange,
     loadNextPage,
-  } = useSearchPlaces();
-
-  // 검색 결과 선택 시 호출되는 함수
-  const handleSelectPlace = useCallback(
-    (place: PlaceSearchResult) => {
-      createAddressMutation.mutate(place, {
-        onSuccess: () => {
-          console.log('주소 이력 생성 성공 후 refetch 호출');
-          // 성공 후 데이터 다시 조회
-          setTimeout(() => {
-            console.log('refetch 호출');
-            // 캐시 무효화 후 다시 조회
-            queryClient.invalidateQueries({ queryKey: ['addressHistory', 5, sort] });
-            refetch();
-          }, 500); // 더 긴 지연을 두어 서버 데이터가 준비되도록 함
-        },
-      });
-
-      // 선택한 장소 정보를 URL 파라미터로 전달하여 rentalPage로 이동
-      const searchParams = new URLSearchParams({
-        lat: place.y.toString(),
-        lng: place.x.toString(),
-        address: place.road_address_name || place.address_name,
-        placeName: place.place_name,
-      });
-
-      router.push(`/rental?${searchParams.toString()}`);
-    },
-    [createAddressMutation, refetch, sort, queryClient, router],
-  );
-
-  // 주소 이력 클릭 시 호출되는 함수
-  const handleAddressHistoryClick = useCallback(
-    async (item: { addressId: number; address_name: string }) => {
-      setKeyword(item.address_name);
-
-      // 낙관적 업데이트: 클릭한 주소를 맨 위로 이동
-      const queryKey = ['addressHistory', 5, sort];
-      const currentData = queryClient.getQueryData(queryKey);
-
-      if (currentData && typeof currentData === 'object' && 'pages' in currentData) {
-        const oldData = currentData as {
-          pages: Array<{
-            content?: {
-              getAddressResponses?: Array<{
-                addressId: number;
-                address_name: string;
-                id: string;
-                phone: string;
-                place_name: string;
-                road_address_name: string;
-                x: number;
-                y: number;
-                lastUsed?: number;
-              }>;
-            };
-          }>;
-        };
-
-        // 클릭한 주소를 찾아서 맨 위로 이동
-        const updatedPages = oldData.pages.map((page) => {
-          if (page?.content?.getAddressResponses) {
-            const responses = [...page.content.getAddressResponses];
-            const clickedIndex = responses.findIndex((resp) => resp.addressId === item.addressId);
-
-            if (clickedIndex !== -1) {
-              // 클릭한 주소를 맨 위로 이동하고 lastUsed 업데이트
-              const clickedItem = { ...responses[clickedIndex], lastUsed: Date.now() };
-              responses.splice(clickedIndex, 1);
-              responses.unshift(clickedItem);
-
-              return {
-                ...page,
-                content: {
-                  ...page.content,
-                  getAddressResponses: responses,
-                },
-              };
-            }
-          }
-          return page;
-        });
-
-        // 낙관적 업데이트 적용
-        queryClient.setQueryData(queryKey, {
-          ...oldData,
-          pages: updatedPages,
-        });
-      }
-
-      try {
-        await updateAddressUsageTime(item.addressId);
-      } catch (error) {
-        console.error('주소 사용 시간 업데이트 실패:', error);
-        // 에러 시 캐시 무효화하여 서버 데이터로 복원
-        queryClient.invalidateQueries({ queryKey });
-      }
-    },
-    [setKeyword, queryClient, sort],
-  );
-
-  // 주소 삭제 핸들러
-  const handleDeleteAddress = useCallback(
-    (addressId: number, e: React.MouseEvent) => {
-      e.stopPropagation();
-      deleteAddressMutation.mutate(addressId, {
-        onSuccess: () => {
-          // 성공 후 데이터 다시 조회
-          refetch();
-        },
-      });
-    },
-    [deleteAddressMutation, refetch],
-  );
-
-  // 키워드 변경 핸들러
-  const handleKeywordChange = useCallback(
-    (value: string) => {
-      setKeyword(value);
-    },
-    [setKeyword],
-  );
-
-  // 스크롤 핸들러
-  const handleScroll = useCallback(() => {
-    if (!scrollRef.current) return;
-
-    const { scrollTop, scrollHeight, clientHeight } = scrollRef.current;
-    const isNearBottom = scrollTop + clientHeight >= scrollHeight - 100;
-
-    if (isNearBottom && hasNextPage && !isFetchingNextPage) {
-      fetchNextPage();
-    }
-  }, [fetchNextPage, hasNextPage, isFetchingNextPage]);
-
-  const throttledHandleScroll = useThrottledScroll(handleScroll, 100);
-
-  // 스크롤 이벤트 리스너
-  useEffect(() => {
-    const scrollElement = scrollRef.current;
-    if (scrollElement) {
-      scrollElement.addEventListener('scroll', throttledHandleScroll);
-      return () => scrollElement.removeEventListener('scroll', throttledHandleScroll);
-    }
-  }, [throttledHandleScroll]);
-
-  // 주소 이력 존재 여부
-  const hasAddressHistory = useMemo(() => {
-    return (
-      Array.isArray(addressHistoryInfinite?.pages) &&
-      addressHistoryInfinite.pages.some(
-        (page) =>
-          Array.isArray(page?.content?.getAddressResponses) &&
-          page.content.getAddressResponses.length > 0,
-      )
-    );
-  }, [addressHistoryInfinite?.pages]);
+    scrollRef,
+  } = useSearchPos();
 
   return (
     <BaseLayout
@@ -223,22 +75,26 @@ export default function SearchPosPage() {
       <div className="w-full h-full flex flex-col">
         {/* 고정 영역 - 검색창 */}
         <div className="flex-shrink-0">
-          <SearchInputField keyword={keyword} onKeywordChange={handleKeywordChange} />
+          <Suspense fallback={<SearchInputLoadingFallback />}>
+            <SearchInputField keyword={keyword} onKeywordChange={handleKeywordChange} />
+          </Suspense>
         </div>
 
         {/* 스크롤 영역 - 목록만 (검색창 아래 여백 추가) */}
-        <div ref={scrollRef} className="flex-1 overflow-y-auto custom-scrollbar pt-20">
+        <div ref={scrollRef} className="flex-1 overflow-y-auto custom-scrollbar pt-5">
           {keyword.trim() ? (
             // 검색어가 있을 때 - 검색 결과 표시
-            <SearchResults
-              results={searchResults}
-              isLoading={isSearchLoading}
-              isLoadingMore={isLoadingMore}
-              hasNext={hasNext}
-              onSelectPlace={handleSelectPlace}
-              onLoadMore={loadNextPage}
-              keyword={keyword.trim()}
-            />
+            <Suspense fallback={<SearchResultsLoadingFallback />}>
+              <SearchResults
+                results={searchResults}
+                isLoading={isSearchLoading}
+                isLoadingMore={isLoadingMore}
+                hasNext={hasNext}
+                onSelectPlace={handleSelectPlace}
+                onLoadMore={loadNextPage}
+                keyword={keyword.trim()}
+              />
+            </Suspense>
           ) : (
             // 검색어가 없을 때 - 현재 위치 버튼과 최근 검색 주소 표시
             <>
@@ -248,12 +104,14 @@ export default function SearchPosPage() {
               )}
 
               {/* 주소 이력 목록 */}
-              <AddressHistoryList
-                addressHistoryInfinite={addressHistoryInfinite ?? { pages: [] }}
-                onAddressClick={handleAddressHistoryClick}
-                onDelete={handleDeleteAddress}
-                isFetchingNextPage={isFetchingNextPage}
-              />
+              <Suspense fallback={<AddressHistoryLoadingFallback />}>
+                <AddressHistoryList
+                  addressHistoryInfinite={addressHistoryInfinite ?? { pages: [] }}
+                  onAddressClick={handleAddressHistoryClick}
+                  onDelete={handleDeleteAddress}
+                  isFetchingNextPage={isFetchingNextPage}
+                />
+              </Suspense>
             </>
           )}
         </div>
