+++ src/features/rental/search/ui/AddressHistoryItem.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback } from 'react';
+import React, { useCallback, useState } from 'react';
 
 import { MapPin, Phone, X } from 'lucide-react';
 
@@ -19,6 +19,8 @@ const AddressHistoryItem = React.memo(
     onAddressClick: (item: { addressId: number; address_name: string }) => void;
     onDelete: (addressId: number, e: React.MouseEvent) => void;
   }) => {
+    const [isDeleting, setIsDeleting] = useState(false);
+
     const handleClick = useCallback(() => {
       onAddressClick(item);
     }, [item, onAddressClick]);
@@ -26,14 +28,21 @@ const AddressHistoryItem = React.memo(
     const handleDelete = useCallback(
       (e: React.MouseEvent) => {
         e.preventDefault();
-        onDelete(item.addressId, e);
+        setIsDeleting(true);
+
+        // 애니메이션 완료 후 삭제 실행
+        setTimeout(() => {
+          onDelete(item.addressId, e);
+        }, 300);
       },
       [item.addressId, onDelete],
     );
 
     return (
       <div
-        className="p-3 border-b border-[var(--gray-light)] cursor-pointer hover:bg-[var(--gray-light)] transition-colors"
+        className={`p-3 border-b border-[var(--gray-light)] rounded-lg cursor-pointer hover:bg-[var(--gray-light)] transition-all duration-300 ${
+          isDeleting ? 'animate-slide-out-down opacity-0' : ''
+        }`}
         onClick={handleClick}
       >
         <div className="flex items-start gap-3">
@@ -66,9 +75,9 @@ const AddressHistoryItem = React.memo(
           </div>
           <button
             onClick={handleDelete}
-            className="p-1 hover:bg-[var(--gray-light)] rounded transition-colors flex-shrink-0"
+            className="cursor-pointer p-1 rounded transition-colors flex-shrink-0"
           >
-            <X className="w-3 h-3 text-[var(--gray)]" />
+            <X className="w-3 h-3 text-[var(--gray)] hover:text-[var(--gray-dark)] " />
           </button>
         </div>
       </div>
